Updated OR function to handle away state.
Note that in the event that an away probe starts with both TLS and PRB asserted, both will need to de-assert in order to register as probed (this is the nature of XOR logic function).  This will still work in the event that the tool probe is probed with the touch probe.  But a misconfigured system might not work as expected (ie, if tool probe polarity is incorrect at the start of a G38.4 move)